+++ ui/src/Components/SaveButton.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { submit, isPristine } from 'redux-form';
+import { submit, isPristine, isValid } from 'redux-form';
 
 const onClickHandler = (dispatch, formName) => () => {
     dispatch(submit(formName));
 };
 
-const SaveButton = ({ dispatch, formName, className, isPristineForm }) => (
+const SaveButton = ({ dispatch, formName, className, isPristineForm, isValidForm }) => (
     <button
         className={`btn btn-success ${className}`}
         type="button"
-        disabled={isPristineForm}
+        disabled={isPristineForm || !isValidForm}
         onClick={onClickHandler(dispatch, formName)}
         data-testid="save-btn"
     >
@@ -24,13 +24,18 @@ SaveButton.propTypes = {
     formName: PropTypes.string.isRequired,
     className: PropTypes.string,
     isPristineForm: PropTypes.bool,
+    isValidForm: PropTypes.bool,
 };
 
 SaveButton.defaultProps = {
     className: '',
     isPristineForm: false,
+    isValidForm: true,
 };
 
-const mapStateToProps = (state, props) => ({ isPristineForm: isPristine(props.formName)(state) });
+const mapStateToProps = (state, props) => ({
+    isPristineForm: isPristine(props.formName)(state),
+    isValidForm: isValid(props.formName)(state),
+});
 
 export default connect(mapStateToProps)(SaveButton);