+++ ui/src/Containers/AccessControl/Roles/Permissions/Form.js
@@ -1,12 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { reduxForm } from 'redux-form';
+import { connect } from 'react-redux';
+import { reduxForm, getFormValues } from 'redux-form';
 
+import FormFieldRequired from 'Components/forms/FormFieldRequired';
 import ReduxTextField from 'Components/forms/ReduxTextField';
 import PermissionsMatrix from 'Containers/AccessControl/Roles/Permissions/PermissionsMatrix/PermissionsMatrix';
 import { defaultMinimalReadAccessResources } from 'constants/accessControl';
 
-const Form = ({ handleSubmit, initialValues, onSubmit }) => {
+const Form = ({ handleSubmit, initialValues, onSubmit, values }) => {
     const disableNameField = !!initialValues && !!initialValues.name;
     return (
         <form
@@ -16,7 +18,9 @@ const Form = ({ handleSubmit, initialValues, onSubmit }) => {
         >
             <div className="mb-4 flex flex-wrap md:flex-no-wrap items-center">
                 <div className="flex-shrink-0 w-full md:w-1/3 pr-8 mb-4 md:mb-0">
-                    <div className="py-2 text-base-600 font-700 text-lg">Role Name</div>
+                    <div className="py-2 text-base-600 font-700 text-lg">
+                        Role Name <FormFieldRequired empty={!values.name.trim()} />
+                    </div>
                     <div data-testid="role-name" className="pb-2">
                         <ReduxTextField name="name" disabled={disableNameField} />
                     </div>
@@ -49,13 +53,33 @@ Form.propTypes = {
         name: PropTypes.string,
         resourceToAccess: PropTypes.shape({}),
     }),
+    values: PropTypes.shape({
+        name: PropTypes.string,
+    }).isRequired,
 };
 
 Form.defaultProps = {
     initialValues: null,
 };
 
+const validate = (values) => {
+    const errors = {};
+
+    if (!values.name.trim()) {
+        errors.name = 'This field is required';
+    }
+
+    return errors;
+};
+
+const form = 'role-form';
+
+const mapStateToProps = (state) => ({
+    values: getFormValues(form)(state),
+});
+
 export default reduxForm({
-    form: 'role-form',
+    form,
     enableReinitialize: true,
-})(Form);
+    validate,
+})(connect(mapStateToProps)(Form));